thoi-khoa-bieu-parser/requirements.txt
openpyxl==3.0.9
python-dotenv==0.19.2
pika==1.3.1